✨ 추천 로직 보완 - ProductDTO, ProductVO에 모험성향과 재정체력 타입을 Integer -> Dobule로 변경 - valueTag, speedTags의 각 ENUM 값을 점수로 매핑해서 거리값 사용 - strategyTag 는 기존과 동일. 단, 0 과 1이 바뀌어있던 버그 수정

+++ src/main/java/org/finmate/product/domain/ProductVO.java
@@ -26,11 +26,11 @@ public class ProductVO {
     private String description; //상품 소개 요약
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
-    private Integer adventureScore;
+    private Double adventureScore;
     private String valueTag;
     private String speedTag;
     private String strategyTag;
-    private Integer minFinanceScore;
+    private Double minFinanceScore;
 
     private DepositVO deposit;
     private SavingsVO savings;
